--- crypto/cross_sign.go
@@ -1,384 +1,0 @@
-// Copyright (c) 2020 Nikos Filippakis
-//
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-package crypto
-
-import (
-	"errors"
-	"fmt"
-
-	"maunium.net/go/mautrix"
-	"maunium.net/go/mautrix/crypto/olm"
-	"maunium.net/go/mautrix/id"
-)
-
-var (
-	ErrCrossSigningKeysNotCached = errors.New("cross-signing private keys not in cache")
-	ErrUserSigningKeyNotCached = errors.New("user-signing private key not in cache")
-	ErrSelfSigningKeyNotCached = errors.New("self-signing private key not in cache")
-	ErrOlmAccountNotLoaded       = errors.New("olm account has not been loaded")
-	ErrSignatureUploadFail       = errors.New("server-side failure uploading signatures")
-	ErrUserNotInQueryResponse    = errors.New("could not find user in query keys response")
-	ErrDeviceNotInQueryResponse  = errors.New("could not find device in query keys response")
-)
-
-func (mach *OlmMachine) storeCrossSigningKeys(crossSigningKeys map[id.UserID]mautrix.CrossSigningKeys, deviceKeys map[id.UserID]map[id.DeviceID]mautrix.DeviceKeys) {
-	for userID, userKeys := range crossSigningKeys {
-		currentKeys, err := mach.CryptoStore.GetCrossSigningKeys(userID)
-		if err != nil {
-			mach.Log.Error("Error fetching current cross-signing keys of user %v: %v", userID, err)
-		}
-		if currentKeys != nil {
-			for curKeyUsage, curKey := range currentKeys {
-				// got a new key with the same usage as an existing key
-				for _, newKeyUsage := range userKeys.Usage {
-					if newKeyUsage == curKeyUsage {
-						if _, ok := userKeys.Keys[id.NewKeyID(id.KeyAlgorithmEd25519, curKey.String())]; !ok {
-							// old key is not in the new key map so we drop signatures made by it
-							if count, err := mach.CryptoStore.DropSignaturesByKey(userID, curKey); err != nil {
-								mach.Log.Error("Error deleting old signatures: %v", err)
-							} else {
-								mach.Log.Debug("Dropped %v signatures made by key `%v` (%v) as it has been replaced", count, curKey, curKeyUsage)
-							}
-						}
-					}
-				}
-			}
-		}
-
-		for _, key := range userKeys.Keys {
-			for _, usage := range userKeys.Usage {
-				mach.Log.Debug("Storing cross-signing key for %v: %v (type %v)", userID, key, usage)
-				if err := mach.CryptoStore.PutCrossSigningKey(userID, usage, key); err != nil {
-					mach.Log.Error("Error storing cross-signing key: %v", err)
-				}
-			}
-
-			for signUserID, keySigs := range userKeys.Signatures {
-				for signKeyID, signature := range keySigs {
-					_, signKeyName := signKeyID.Parse()
-					signingKey := id.Ed25519(signKeyName)
-					// if the signer is one of this user's own devices, find the key from the key ID
-					if signUserID == userID {
-						ownDeviceID := id.DeviceID(signKeyName)
-						if ownDeviceKeys, ok := deviceKeys[userID][ownDeviceID]; ok {
-							signingKey = ownDeviceKeys.Keys.GetEd25519(ownDeviceID)
-							mach.Log.Debug("Treating %v as the device name", signKeyName)
-						}
-					}
-
-					mach.Log.Debug("Verifying with key %v of user %v", signingKey, signUserID)
-					if verified, err := olm.VerifySignatureJSON(userKeys, signUserID, signKeyName, signingKey); err != nil {
-						mach.Log.Error("Error while verifying cross-signing keys: %v", err)
-					} else {
-						if verified {
-							mach.Log.Debug("Cross-signing keys verified")
-							mach.CryptoStore.PutSignature(userID, key, signUserID, signingKey, signature)
-						} else {
-							mach.Log.Error("Cross-signing keys verification unsuccessful", err)
-						}
-					}
-				}
-			}
-		}
-	}
-}
-
-// IsDeviceTrusted returns whether a device has been determined to be trusted either through verification or cross-signing.
-func (mach *OlmMachine) IsDeviceTrusted(device *DeviceIdentity) bool {
-	userID := device.UserID
-	if device.Trust == TrustStateVerified {
-		return true
-	} else if device.Trust == TrustStateBlacklisted {
-		return false
-	}
-	if !mach.IsUserTrusted(userID) {
-		return false
-	}
-
-	theirKeys, err := mach.CryptoStore.GetCrossSigningKeys(userID)
-	if err != nil {
-		mach.Log.Error("Error retrieving cross-singing key of user %v from database: %v", userID, err)
-		return false
-	}
-	theirMSK, ok := theirKeys[id.XSUsageMaster]
-	if !ok {
-		mach.Log.Error("Master key of user %v not found", userID)
-		return false
-	}
-	theirSSK, ok := theirKeys[id.XSUsageSelfSigning]
-	if !ok {
-		mach.Log.Error("Self-signing key of user %v not found", userID)
-		return false
-	}
-	sskSigExists, err := mach.CryptoStore.IsKeySignedBy(userID, theirSSK, userID, theirMSK)
-	if err != nil {
-		mach.Log.Error("Error retrieving cross-singing signatures for master key of user %v from database: %v", userID, err)
-		return false
-	}
-	if !sskSigExists {
-		mach.Log.Warn("Self-signing key of user %v is not signed by their master key", userID)
-		return false
-	}
-	deviceSigExists, err := mach.CryptoStore.IsKeySignedBy(userID, device.SigningKey, userID, theirSSK)
-	if err != nil {
-		mach.Log.Error("Error retrieving cross-singing signatures for master key of user %v from database: %v", userID, err)
-		return false
-	}
-	return deviceSigExists
-}
-
-// IsUserTrusted returns whether a user has been determined to be trusted by our user-signing key having signed their master key.
-// In the case the user ID is our own and we have successfully retrieved our cross-signing keys, we trust our own user.
-func (mach *OlmMachine) IsUserTrusted(userID id.UserID) bool {
-	if mach.crossSigningKeys == nil {
-		return false
-	}
-	if userID == mach.Client.UserID {
-		return true
-	}
-	// first we verify our user-signing key
-	sskSigs, err := mach.CryptoStore.GetSignaturesForKeyBy(mach.Client.UserID, mach.crossSigningKeys.UserSigningKey.PublicKey, mach.Client.UserID)
-	if err != nil {
-		mach.Log.Error("Error retrieving our self-singing key signatures: %v", err)
-		return false
-	}
-	if _, ok := sskSigs[mach.crossSigningKeys.MasterKey.PublicKey]; !ok {
-		// our user-signing key was not signed by our master key
-		return false
-	}
-	theirKeys, err := mach.CryptoStore.GetCrossSigningKeys(userID)
-	if err != nil {
-		mach.Log.Error("Error retrieving cross-singing key of user %v from database: %v", userID, err)
-		return false
-	}
-	theirMskKey, ok := theirKeys[id.XSUsageMaster]
-	if !ok {
-		mach.Log.Error("Master key of user %v not found", userID)
-		return false
-	}
-	sigExists, err := mach.CryptoStore.IsKeySignedBy(userID, theirMskKey, mach.Client.UserID, mach.crossSigningKeys.UserSigningKey.PublicKey)
-	if err != nil {
-		mach.Log.Error("Error retrieving cross-singing signatures for master key of user %v from database: %v", userID, err)
-		return false
-	}
-	return sigExists
-}
-
-// UploadSignedCrossSigningKeys signs and uploads the given cached cross-signing keys to the server.
-func (mach *OlmMachine) UploadSignedCrossSigningKeys(keys *CrossSigningKeysCache, uiaCallback mautrix.UIACallback) error {
-	userID := mach.Client.UserID
-	masterKeyID := id.NewKeyID(id.KeyAlgorithmEd25519, keys.MasterKey.PublicKey.String())
-	masterKey := mautrix.CrossSigningKeys{
-		UserID: userID,
-		Usage:  []id.CrossSigningUsage{id.XSUsageMaster},
-		Keys: map[id.KeyID]id.Ed25519{
-			masterKeyID: keys.MasterKey.PublicKey,
-		},
-	}
-
-	selfKey := mautrix.CrossSigningKeys{
-		UserID: userID,
-		Usage:  []id.CrossSigningUsage{id.XSUsageSelfSigning},
-		Keys: map[id.KeyID]id.Ed25519{
-			id.NewKeyID(id.KeyAlgorithmEd25519, keys.SelfSigningKey.PublicKey.String()): keys.SelfSigningKey.PublicKey,
-		},
-	}
-	selfSig, err := keys.MasterKey.SignJSON(selfKey)
-	if err != nil {
-		return fmt.Errorf("failed to sign self-signing key: %w", err)
-	}
-	selfKey.Signatures = map[id.UserID]map[id.KeyID]string{
-		userID: {
-			masterKeyID: selfSig,
-		},
-	}
-
-	userKey := mautrix.CrossSigningKeys{
-		UserID: userID,
-		Usage:  []id.CrossSigningUsage{id.XSUsageUserSigning},
-		Keys: map[id.KeyID]id.Ed25519{
-			id.NewKeyID(id.KeyAlgorithmEd25519, keys.UserSigningKey.PublicKey.String()): keys.UserSigningKey.PublicKey,
-		},
-	}
-	userSig, err := keys.MasterKey.SignJSON(userKey)
-	if err != nil {
-		return fmt.Errorf("failed to sign user-signing key: %w", err)
-	}
-	userKey.Signatures = map[id.UserID]map[id.KeyID]string{
-		userID: {
-			masterKeyID: userSig,
-		},
-	}
-
-	return mach.Client.UploadCrossSigningKeys(&mautrix.UploadCrossSigningKeysReq{
-		Master:      masterKey,
-		SelfSigning: selfKey,
-		UserSigning: userKey,
-	}, uiaCallback)
-}
-
-// SignUserAndUpload creates a cross-signing signature for a user, stores it and uploads it to the server.
-func (mach *OlmMachine) SignUserAndUpload(userID id.UserID, masterKey id.Ed25519) error {
-	if userID == mach.Client.UserID {
-		return nil
-	} else if mach.crossSigningKeys == nil || mach.crossSigningKeys.UserSigningKey == nil {
-		return ErrUserSigningKeyNotCached
-	}
-
-	userSigningKey := mach.crossSigningKeys.UserSigningKey
-	masterKeyObj := mautrix.ReqKeysSignatures{
-		UserID: userID,
-		Usage:  []id.CrossSigningUsage{id.XSUsageMaster},
-		Keys: map[id.KeyID]string{
-			id.NewKeyID(id.KeyAlgorithmEd25519, masterKey.String()): masterKey.String(),
-		},
-	}
-	signature, err := userSigningKey.SignJSON(masterKeyObj)
-	if err != nil {
-		return fmt.Errorf("failed to sign JSON: %w", err)
-	}
-	masterKeyObj.Signatures = mautrix.Signatures{
-		mach.Client.UserID: map[id.KeyID]string{
-			id.NewKeyID(id.KeyAlgorithmEd25519, userSigningKey.PublicKey.String()): signature,
-		},
-	}
-	mach.Log.Trace("Signed master key for user %v: `%v`", userID, signature)
-
-	resp, err := mach.Client.UploadSignatures(&mautrix.ReqUploadSignatures{
-		userID: map[string]mautrix.ReqKeysSignatures{
-			masterKey.String(): masterKeyObj,
-		},
-	})
-
-	if err != nil {
-		return fmt.Errorf("error while uploading signatures: %w", err)
-	} else if len(resp.Failures) > 0 {
-		return fmt.Errorf("%w: %+v", ErrSignatureUploadFail, resp.Failures)
-	}
-
-	if err := mach.CryptoStore.PutSignature(userID, masterKey, mach.Client.UserID, userSigningKey.PublicKey, signature); err != nil {
-		return fmt.Errorf("error storing signature in crypto store: %w", err)
-	}
-
-	return nil
-}
-
-// SignOwnMasterKeyAndUpload uses the current account for signing the current user's master key and uploads the signature.
-func (mach *OlmMachine) SignOwnMasterKeyAndUpload() error {
-	if mach.crossSigningKeys == nil {
-		return ErrCrossSigningKeysNotCached
-	} else if mach.account == nil {
-		return ErrOlmAccountNotLoaded
-	}
-
-	userID := mach.Client.UserID
-	deviceID := mach.Client.DeviceID
-	masterKey := mach.crossSigningKeys.MasterKey.PublicKey
-
-	masterKeyObj := mautrix.ReqKeysSignatures{
-		UserID: userID,
-		Usage:  []id.CrossSigningUsage{id.XSUsageMaster},
-		Keys: map[id.KeyID]string{
-			id.NewKeyID(id.KeyAlgorithmEd25519, masterKey.String()): masterKey.String(),
-		},
-	}
-	signature, err := mach.account.Internal.SignJSON(masterKeyObj)
-	if err != nil {
-		return fmt.Errorf("failed to sign JSON: %w", err)
-	}
-	masterKeyObj.Signatures = mautrix.Signatures{
-		userID: map[id.KeyID]string{
-			id.NewKeyID(id.KeyAlgorithmEd25519, deviceID.String()): signature,
-		},
-	}
-	mach.Log.Trace("Signed own master key with device %v: `%v`", deviceID, signature)
-
-	resp, err := mach.Client.UploadSignatures(&mautrix.ReqUploadSignatures{
-		userID: map[string]mautrix.ReqKeysSignatures{
-			masterKey.String(): masterKeyObj,
-		},
-	})
-
-	if err != nil {
-		return fmt.Errorf("error while uploading signatures: %w", err)
-	} else if len(resp.Failures) > 0 {
-		return fmt.Errorf("%w: %+v", ErrSignatureUploadFail, resp.Failures)
-	}
-
-	if err := mach.CryptoStore.PutSignature(userID, masterKey, userID, mach.account.SigningKey(), signature); err != nil {
-		return fmt.Errorf("error storing signature in crypto store: %w", err)
-	}
-
-	return nil
-}
-
-// SignOwnDeviceAndUpload creates a cross-signing signature for a device belonging to the current user and uploads it to the server.
-func (mach *OlmMachine) SignOwnDeviceAndUpload(deviceID id.DeviceID) error {
-	if mach.crossSigningKeys == nil || mach.crossSigningKeys.SelfSigningKey == nil {
-		return ErrSelfSigningKeyNotCached
-	}
-
-	userID := mach.Client.UserID
-	selfSigningKey := mach.crossSigningKeys.SelfSigningKey
-
-	devicesKeys, err := mach.Client.QueryKeys(&mautrix.ReqQueryKeys{
-		DeviceKeys: mautrix.DeviceKeysRequest{
-			userID: mautrix.DeviceIDList{deviceID},
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("error querying device keys for %s: %w", deviceID, err)
-	}
-	userKeys, ok := devicesKeys.DeviceKeys[userID]
-	if !ok {
-		return ErrUserNotInQueryResponse
-	}
-	deviceObj, ok := userKeys[deviceID]
-	if !ok {
-		return ErrDeviceNotInQueryResponse
-	}
-
-	deviceKeyObj := mautrix.ReqKeysSignatures{
-		UserID:     userID,
-		DeviceID:   deviceID,
-		Algorithms: deviceObj.Algorithms,
-		Keys:       make(map[id.KeyID]string),
-	}
-	for keyID, key := range deviceObj.Keys {
-		deviceKeyObj.Keys[id.KeyID(keyID)] = key
-	}
-
-	signature, err := selfSigningKey.SignJSON(deviceKeyObj)
-	if err != nil {
-		return fmt.Errorf("failed to sign JSON: %w", err)
-	}
-	deviceKeyObj.Signatures = mautrix.Signatures{
-		userID: map[id.KeyID]string{
-			id.NewKeyID(id.KeyAlgorithmEd25519, selfSigningKey.PublicKey.String()): signature,
-		},
-	}
-
-	mach.Log.Trace("Signed own device %v with self-signing key: `%v`", deviceID, signature)
-
-	resp, err := mach.Client.UploadSignatures(&mautrix.ReqUploadSignatures{
-		userID: map[string]mautrix.ReqKeysSignatures{
-			deviceID.String(): deviceKeyObj,
-		},
-	})
-
-	if err != nil {
-		return fmt.Errorf("error while uploading signatures: %w", err)
-	} else if len(resp.Failures) > 0 {
-		return fmt.Errorf("%w: %+v", ErrSignatureUploadFail, resp.Failures)
-	}
-
-	if err := mach.CryptoStore.PutSignature(userID, deviceObj.Keys.GetEd25519(deviceID), userID, selfSigningKey.PublicKey, signature); err != nil {
-		return fmt.Errorf("error storing signature in crypto store: %w", err)
-	}
-
-	return nil
-}
